pyproject.toml
[project]
name = "imagetranslaterai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "김동환", email = "kimdonghwandeveloper@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "openai>=2.15.0",
    "opencv-python>=4.12.0.88",
    "paddleocr>=3.3.2",
    "paddlepaddle>=3.2.2",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
imagetranslaterai = "imagetranslaterai:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
